is: Use `ON DELETE CASCADE` foreign key constraint
Previous CockroachDB versions did not include this feature so
deleting in the SQL store rows that were being referenced in other
rows involved first to delete these last mentioned rows.

As we just upgraded to the newest CockroachDB (2.0.0) which includes
the `ON DELETE CASCADE` foreign constraint we do not need to delete
anymore manually those rows that references the parent row.