util: Fix ping slot channel in band
Default ping slot channel is defined for some bands and for some it's not. Existing field is less than useful, given that it stores ping slot channel in a slice and for bands, where there's no ping slot channel defined, it just sets it to list of fixed channels for some mysterious reason.
Fix that by changing the type of field to *uint64. If the default ping slot frequency is defined by band, the field will be set to pointer to the value, otherwise it's <nil>.